Refactor to allow passing filtering/limiting args to each method
Given the long goal of filtering/limiting across states, remove
state as a required argument to the constructor and accept
filtering arguments to the different methods, imagining an
eventual chainable interface that breaks out filtering, limiting
and writing as separate calls.

I don't know if this is the best way to do it, but it makes the
most sense to me today.

Addresses #39