Add (raw_)party field to Result, remove from Candidate
Explicitely add ``party`` and ``raw_party`` fields to the ``Result``
model.  This accomodates the case where a candidate can be the
nominee from multiple parties in states like NY and CT, and reflects
the way that the raw data delivers results on a per-party,
per-candidate basis.

In the case of ``us.md.load``, it looks like party was added as a
dynamic field of ``Result`` in most cases anyway.

Remove these fields from ``Candidate``.

Addresses #46