Allow association of validators with transforms
Create a ``Transform`` class to wrap the transform function and
aggregate related validators.

Update ``Registry.register()`` to accept a list of validators.

Return lists when fetching validators and transforms rather than
dicts.  Use func.__name__ to get the name.

Factor out the actual running of validation into
``tasks.validate.run_validation`` so it can be reused to run
validations with the same output in places other than the
``validate.run`` task.

Update the ``transform.run`` task to run validators associated
with a transform.

Update the ``transform.list`` task to list validators associated
with a transform.

Addresses #62