Contest should only have one party field
Previous Contest had ``party`` and ``primary_party`` although
the help text for the fields indicates that they mean the same
thing.  There should only be one field.  I decided that it should
be a ``ReferenceField`` to ``Party`` and that it should be called
``primary_party`` as it has the same meaning as that field on
``RawResult``.

Addresses #49