Automatic timestamps and Contest slugs
Make the common timestamp fields on many models more DRY by putting them
in a mixin. We have to use a mixin rather than a base class because
MongoEngine doesn't let you inherit from a Document class without
affecting the collection where the document is stored in Mongo.

Adds a dependency of ``blinker`` to enable MongoEngine signals.

Use a signal handler to automatically set the slug field when a
Contest is constucted.

Addresses #49