s/_build_meta_fields/_build_common_election_kwargs/
Rename MDBaseLoader._build_meta_fields() to
MDBaseLoader._build_common_election_kwargs().

This is because we're ultimately building keyword arguments that
will being passed to the RawResult constructor.  Also, the _kwargs
naming convention is used in other similar cases.

I chose not to use the term "meta" because the keyword arguments
we're building include some information that will be in contest
records.  They're different than the fields labeled as meta in
models.  Ultimately, what they share is that they're common to a
particular election.

Addresses #49