Refinements to Maryland transforms
Factor out buffered bulk mongo inserts

There are too many result records created in the
create_unique_results transform for a single call to
Result.objects.insert().  Create a class that inserts batches
of records at a time.

Use a non-caching queryset for RawResults

We have to use a non-caching queryset when iterating through
all the RawResults.  Otherwise, we run out of memory.

Remove the source field when looking up candidates

A candidate may be present in multiple source files, so this
shouldn't be used as a lookup field.

Addresses #49