Move generally useful methods to BaseLoader
Move methods from Maryland's base loader class to ``BaseLoader``.

In starting to think about the Florida loader, it's clear that
the functionality provided in these methods would be duplicated,
so it makes sense to just move them to the base class.

This makes the hook that should be implemented in state-specific
sublcasses the ``load()`` method rather than ``run()``.

Add docstrings to make the purpose of these methods more clear
and to help contributors understand how to use or implement the
methods in their subclasses.