Validators for result, contest and candidate counts
Add validators for result, contest and candidate counts for
Maryland's 2000 primary.  In order to describe and calculate the
expected counts, I created two utility classes,
``MDElectionDescription`` and ``Primary2000ElectionDescription``.

This was somewhat tedious to implement, but organizes this information
in a better way than calculating it within the validator functions.

It might make sense to move these definitions, especially the lists
of counties and districts, to a separate module/package or add them
to the metadata collection process.

Addresses https://github.com/openelections/core/issues/36