Standardize reversal of transforms
In the old code that I used as the base for some of the new
Maryland transforms, old records were deleted first.  This was
the case for creating Result records, but not for Contest or
Candidate records.  In my first pass at class-based transforms,
I created a ``reverse()`` hook to break out the code that reverses
the transform, but not a clear way to express whether this should
be automatically run before the transform is run.

I added a ``auto_reverse`` attribute to the base transform class.
If True, the reverse will be run automatically before the transform
is run, each time it is invoked.  This can be supressed using the
--no-reverse flag to the ``transform.run`` task.

I added a separate ``transform.reverse`` task to provide a way to
easily reverse a transform.  This is particularly helpful in
testing and development of transforms.