Add Maryland 2012 election validators
The 2010 redistricting process changed the county to congressional
district mappings, so I had to define the new mapping in
``us.md.jurisdiction``.  In addition, I tried to standardize the
naming conventions of the mappings variables in the jurisdiction
module.  I also made the various jurisdiction definitions class
attributes of the election classes to make access within methods
the same.  If a mapping varies between elections, it can be specified
in the class definition.

Addresses https://github.com/openelections/core/issues/36