Custom loader to extract zipped Arkansas results
This a checkpoint commit for implementing the Arkansas datasource
and loader that adds support for extracting the results from a
single zipped archive.

Currently only results for elections in 2000 are supported.

The datasource generates a mapping for each result file (including
ones archived in zip files) and adds a field to the mapping with
the archived file name.

For the general election, it's we can programatically generate the
filenames from the county names.  For the special elections, just
look things up from url_paths.csv

Addresses https://github.com/openelections/core/issues/124