Use generated Arkansas url_paths to repo
These are generated by scraping the Clarity system for certain
elections since, as far as I can tell, there isn't a way to
predict the identifiers in the URLs for the county-level election
summary pages and results.

Though I wanted to generate these from code to document and be
able to replicate my process, I think it makes sense to include
the output in the repo to save others the cost/time of rescraping
the information in their instances.

Addresses https://github.com/openelections/core/issues/124