Add preprocessed URL for Washington .mdb results
Add a preprocessed URL for Washington State results that are
shipped in an Access database to the datasource mappings output.

The actual preprocessed result file still needs to be created
(see https://github.com/openelections/core/issues/150), but the
datasource and fetch task will be ready to consume it once
the preprocessed files are created.

Addresses https://github.com/openelections/core/issues/145