Fix baker to address failing tests
Some tests were failing so I fixed bugs in the code that
were causing the failures.

In Roller.build_filters, reporting_level should not be a
field in the common filters.

The tests for the filename() and manifest_filename() methods
needed to be updated to reflect the updated signatures for those
methods since it was changed to a class method.