Add ocd_id field to RawResult
Add an ``ocd_id`` field to RawResult and output in baked results as
the ``division`` field.

While we originally imagined that creating the ocd_id would happen
in a transform, we need to set it on raw results.  Otherwise, the
baked output records don't indicate the relationship of precincts
to counties (or other parent geographies) for county-level results.

Update loaders and transforms for the following states to add
ocd_id fields to the create RawResult models:

* Florida
* Maryland
* West Virginia
* Wyoming

These are all the states that have implemented loaders thus far.

I made minor changes to Ohio's loader, but it seemed like a work
in progress and it contains copied and pasted code from Maryland.

Addresses https://github.com/openelections/core/issues/190