Checkpoint commit for Iowa Loader
Implement the scaffolding for the Iowa loader including the base
classes for both the preprocessed and Excel results.

Update the mappings to skip loading of some precinct-level results
from a single statewide file.  This appears to be the same information
as the per-county precinct files, and doesn't have candidate names
or district numbers.

Still to do is to finish implementing loading of post-2010,
precinct-level results.

Addresses https://github.com/openelections/core/issues/203