s/BaseLoader._xls_file_handle/BaseLoader._xls_file_path/
Rename the ``BaseLoader._xls_file_handle`` method to
``BaseLoader._xls_file_path`` since it returns a string containing
the path to the Excel file and not a file handle as is the case
with ``BaseLoader._file_handle``.  The difference is because of
the needs of the xlrd API.

Make ``BaseLoader._xls_file_path`` a property like
``BaseLoader._file_handle``.

Update the Wyoming loader to use the new name.

Fixes https://github.com/openelections/core/issues/207