Update installation instructions
Update installation instructions in the README so they instruct the
user to install the ``openelex`` package as an editable package
by running:

pip install -e .

This will automatically put the package on the virtualenv's
PYTHONPATH and create the ``openelex`` executable, so instructions to
do that explicitly are no longer needed.

Remove the ``bin/openelex`` wrapper script since we're now using
setuptools to create the executable.

Addresses https://github.com/openelections/core/issues/227