Insert RawResults in batches in PA loader
Some of the PA files can be quite large, and if we wait until we've
constructed RawResults for each record in the data, we run out of
data when running the load task on prod-1.

Instead, use BulkInsertBuffer to do inserts in batches of 1000.

Addresses https://github.com/openelections/openelections-core/issues/240